+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - #2725 Update webpack build system
 - #2719 jQuery3 compatibility
+- #2719 jQuery3 compatibility
+- #2724 Allow to select an analyst and template in the create worksheet modal
 - #2723 Add registry to control import file attachment to Worksheet assigned analyses
 - #2722 Fix instrument import loggings
 - #2720 Allow to set custom IDs to reference samples